nodejs/desafio-3/api-find-friend/src/http/controllers/organizations/refresh.ts
import { FastifyReply, FastifyRequest } from 'fastify'

export async function refresh(request: FastifyRequest, reply: FastifyReply) {
  await request.jwtVerify({ onlyCookie: true }).catch(() => {
    return reply.status(401).send({ message: 'Unauthorized' })
  })

  const { sub } = request.user

  const token = await reply.jwtSign({}, { sign: { sub } })

  const refreshToken = await reply.jwtSign(
    {},
    { sign: { sub, expiresIn: '7d' } },
  )

  return reply
    .setCookie('refreshToken', refreshToken, {
      path: '/',
      secure: true,
      sameSite: true,
      httpOnly: true,
    })
    .status(200)
    .send({ token })
}
